backend/src/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
dnspython==2.4.1
email-validator==2.0.0.post2
fastapi==0.100.1
fastapi-users==12.1.0
fastapi-users-db-sqlalchemy==6.0.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
